Move SocketSniffer to its own module
