+++ talpid-tunnel-config-client/src/lib.rs
@@ -18,6 +18,9 @@ mod ml_kem;
 #[cfg(not(target_os = "ios"))]
 mod socket;
 
+#[cfg(not(target_os = "ios"))]
+mod socket_sniffer;
+
 #[allow(clippy::derive_partial_eq_without_eq)]
 mod proto {
     tonic::include_proto!("ephemeralpeer");
@@ -277,83 +280,11 @@ async fn connect_relay_config_client(ip: Ipv4Addr) -> Result<RelayConfigService,
         .connect_with_connector(service_fn(move |_| async move {
             let sock = socket::TcpSocket::new()?;
             let stream = sock.connect(addr).await?;
-            let sniffer = socket_sniffer::SocketSniffer {
-                s: stream,
-                rx_bytes: 0,
-                tx_bytes: 0,
-                start_time: std::time::Instant::now(),
-            };
+            let sniffer = socket_sniffer::SocketSniffer::new(stream);
             Ok::<_, std::io::Error>(TokioIo::new(sniffer))
         }))
         .await
         .map_err(Error::GrpcConnectError)?;
 
     Ok(RelayConfigService::new(connection))
-}
-
-mod socket_sniffer {
-    pub struct SocketSniffer<S> {
-        pub s: S,
-        pub rx_bytes: usize,
-        pub tx_bytes: usize,
-        pub start_time: std::time::Instant,
-    }
-    use std::{
-        io,
-        pin::Pin,
-        task::{Context, Poll},
-    };
-
-    use tokio::io::AsyncWrite;
-
-    use tokio::io::{AsyncRead, ReadBuf};
-
-    impl<S> Drop for SocketSniffer<S> {
-        fn drop(&mut self) {
-            let duration = self.start_time.elapsed();
-            log::debug!(
-                "Tunnel config client connection ended. RX: {} bytes, TX: {} bytes, duration: {} s",
-                self.rx_bytes,
-                self.tx_bytes,
-                duration.as_secs()
-            );
-        }
-    }
-
-    impl<S: AsyncRead + AsyncWrite + Unpin> AsyncRead for SocketSniffer<S> {
-        fn poll_read(
-            mut self: Pin<&mut Self>,
-            cx: &mut Context<'_>,
-            buf: &mut ReadBuf<'_>,
-        ) -> Poll<io::Result<()>> {
-            let initial_data = buf.filled().len();
-            let bytes = std::task::ready!(Pin::new(&mut self.s).poll_read(cx, buf));
-            if bytes.is_ok() {
-                self.rx_bytes += buf.filled().len().saturating_sub(initial_data);
-            }
-            Poll::Ready(bytes)
-        }
-    }
-
-    impl<S: AsyncRead + AsyncWrite + Unpin> AsyncWrite for SocketSniffer<S> {
-        fn poll_write(
-            mut self: Pin<&mut Self>,
-            cx: &mut Context<'_>,
-            buf: &[u8],
-        ) -> Poll<io::Result<usize>> {
-            let bytes = std::task::ready!(Pin::new(&mut self.s).poll_write(cx, buf));
-            if let Ok(bytes) = bytes {
-                self.tx_bytes += bytes;
-            }
-            Poll::Ready(bytes)
-        }
-
-        fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            Pin::new(&mut self.s).poll_flush(cx)
-        }
-
-        fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            Pin::new(&mut self.s).poll_shutdown(cx)
-        }
-    }
 }
